[InstCombine] Prefer to keep power-of-2 constants when combining ashr exact and slt/ult of a constant
We have flexibility in what constant to use when combining an `ashr
exact` with a slt or ult of a constant, and it's not possible to revisit
this decision later in the compilation pipeline after the `ashr exact`
is removed. Keeping a constant close to power-of-2 (pow2val + 1) should
be now worse than neutral, and in some cases may allow better codegen
later on for targets that can more cheaply generated power of 2 (which
may be selectable if converting back to setle/setge) or near power of 2
constants.

Alive2: <https://alive2.llvm.org/ce/z/anqsyz> and
<https://alive2.llvm.org/ce/z/AL9P3o>