llvm-objdump: ensure a MachO symbol isn't STAB before looking up section.
The section field has been repurposed for some STAB symbol types, and if we
blindly look it up we'll produce an error and terminate. Logic already existed

Existing stabs test had a section that was in range. Unfortunately I don't know
of an easy way to produce stabs entries in LLVM (I thought they died in the 90s
until this came up) so I just binary-edited it to cause a failure on existing
llvm-objdump.