update_test_checks: refactor to use a new GeneralizerInfo object
This change is motivated primarily by the desire to change the list of
"NamelessValue" (not a great name) objects based on the --version passed
to update_test_checks.

In order to be able to use the version number e.g. when building up
regular expressions, but also to just reduce the number of surprises
caused by global state, this change introduces the GeneralizerInfo class
which is meant to hold a bunch of information about how to generalize
check lines.

The GeneralizerInfo must be instantiated by the "driver" script
(update_*_test_checks.py) and is passed around various common functions.

Beware that the GeneralizerInfo objects are *not* stateless. The
NamelessValue objects currently hold mapping information for "global"
values.