Fix loop cache cost to avoid cost of zero for refgroups.
Currently loop cache analysis uses following formula to evaluate cost of an RefGroup for a consecutive memory access: "RefCost=(TripCount*Stride)/CLS" .The cost evaluates to zero when "TripCount*Strid" is smaller than cache-line-size. This could result in wrong cost value for a loop and misleading loopInterchange decision.

This patch fixes the problem by rounding the cost to 1 once this problem happens.